changed HA groups for Perth low and intermediate groups
I realized that I had specified the wrong groups for the
"low" and "intermediate" divergence groups. When I made
the scatterplots I was getting very long branches in the
"low" divergence alignment. This commit show fix  that by
a) specifying the correct groups in the code and b) making
the new alignments